Support rv32e
As defined in RISC-V Unprivileged Spec. 20191213, the only change in RVE
is to reduce the number of integer registers to 16 (x0-x15). This requires
also a separate calling convention that seems a little underspecified.

Somebody on the internet said these are also implied:

- 2 callee saved registers instead of 12
- 32-bit / 4-byte stack alignment instead of 128 bits / 16 bytes

This patch does the following:

- Disable start of program load-to-zero for x16--x31 for RVE, since these do
  not exist on RVE
- Force & assert the stack pointer is 4-byte aligned on RVE
- (WIP: maybe missing on vectored-rt branch) Configure the trap handler
  to only caller-save a0--a5 and t0--t2 instead of a0--a7 and t0--t6 on RVE
- Disable a6--a7 (x16--x17) and t3--t6 (x28--x31) for RVE in the TrapFrame
  struct, since these do not exist on RVE

This may not be all that is required for RVE support, but this is at
least the minimal amount of changes that makes riscv-rt link on RVE,
allowing further testing.